third_party/WebKit/LayoutTests/fast/dom/shadow/content-deep-expected.txt

CONSOLE WARNING: /deep/ combinator is deprecated. See https://www.chromestatus.com/features/6750456638341120 for more details.
::content /deep/ should match deep into distributed subtree.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS getComputedStyle(target).color is "rgb(0, 128, 0)"
PASS successfullyParsed is true

TEST COMPLETE
